refactor: product rating changed to double

+++ src/main/java/com/shop_master_backend/dto/review/ReviewRequestDTO.java
@@ -10,8 +10,8 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @NoArgsConstructor
 public class ReviewRequestDTO {
-    private Integer rating;
-    private String comment;
-    private String productId;
-    private Integer userId;
+	private Double rating;
+	private String comment;
+	private String productId;
+	private Integer userId;
 }
